Rename _CShims -> _SwiftSyntaxCShims
Avoid possible module name collision
